Add components for VS and VS Code icons (#18)
This PR builds upon #17 (which is to be **reviewed and merged before
this one**, cc @jeremy-rifkin) and adds two inline icons for VS and VS
Code.

Markup:
```
<VSCode /> **VS Code** is a text editor
```

Result:

![image](https://github.com/user-attachments/assets/8cd983ef-59bf-4556-b058-4402a1e5e794)

**Notice:**  
Using those icons (and InlineIcon in general) in Markdown causes janky
line breaks in the rendered HTML. A [bug report about
it](https://github.com/vuejs/vitepress/issues/4626) was opened and
accepted, so I guess we can live with it until it is fixed upstream?