Setup page for VSCode + MinGW (#16)
This PR adds a (big) page to configure VS Code for use with MinGW Clang
and associated tools.
Many thanks to [HolyBlackCat](https://github.com/HolyBlackCat) (can't
tag them apparently) for letting us use their [repo of C++
tutorials](https://github.com/HolyBlackCat/cpp-tutorials) as a basis.

The page is _LONG_. But then again there truly is that much stuff to
configure... Of course, feedback is welcome regardless.

The article uses stuff from #18, which is to be **reviewed and merged
beforehand** (cc @jeremy-rifkin).