Refactor Compiler Development Resources (#39)
### Description
This PR updates the formatting of the **compiler-development** section
to use markdown hyperlinks instead of plain text hyperlinks. This update
makes the page displayed look better.

Additionally, I noticed that two file paths were updated in a previous
PR, but the route was never set to the correct directory. When running
the website and clicking on **compiler-development** and
**performance-analysis**, you got a **404 file not found** error
message. This PR fixed this.